requirements.txt
Django>=5.2,<6.0
stripe>=12.0,<15.0
python-dotenv>=1.0,<2.0
dj-database-url>=2.2,<3.0
gunicorn>=22.0,<24.0
whitenoise>=6.7,<7.0
psycopg[binary]>=3.2,<4.0
pytest>=8.2,<9.0
pytest-django>=4.9,<5.0
